backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
azure-common==1.1.28
azure-core==1.31.0
azure-identity==1.19.0
azure-search-documents==11.5.1
azure-storage-blob==12.23.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.1
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
isodate==0.7.2
jiter==0.6.1
msal==1.31.0
msal-extensions==1.2.0
numpy==2.1.2
openai==1.51.2
pandas==2.2.3
portalocker==2.10.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
starlette==0.40.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
